requirements.txt
asgiref==3.8.1
Django==4.2.20
django-debug-toolbar==5.1.0
django-graphql-jwt==0.4.0
djangorestframework==3.16.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
promise==2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
django-filter==25.1
